% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timefitteR_Data_Model.R
\name{clean_param_list}
\alias{clean_param_list}
\title{Clean TimefitteR Parameter Lists}
\usage{
clean_param_list(p, f)
}
\arguments{
\item{p}{A list of timefitteR_Param_List objects}

\item{f}{A factor with at least 1 level}
}
\value{
A list
}
\description{
Clean TimefitteR Parameter Lists
}
