% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-standardise_model_formula.R
\name{.standardise_model_formula}
\alias{.standardise_model_formula}
\title{Standardise a model formula to use canonical \code{x} and \code{t}}
\usage{
.standardise_model_formula(model, x_col, t_col)
}
\description{
Standardise a model formula to use canonical \code{x} and \code{t}
}
\keyword{internal}
