% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_attitude_sequence.R
\name{generate_attitude_sequence}
\alias{generate_attitude_sequence}
\title{Internal Function to generate a sequence of time points in a semi-deterministic way}
\usage{
generate_attitude_sequence(
  id,
  intercept = 0,
  slope = 1,
  curvature = 0,
  noise = 2,
  gender = "Boy",
  ses = "Low",
  from = 0,
  to = 10,
  gender_intercept_offset = 1,
  gender_intercept_noise = 1,
  gender_slope_offset = -0.1,
  gender_slope_noise = 1,
  medium_intercept_offset = 0.5,
  medium_intercept_noise = 1,
  medium_slope_offset = 0.05,
  medium_slope_noise = 1,
  high_intercept_offset = 1.5,
  high_intercept_noise = 1,
  high_slope_offset = 0.075,
  high_slope_noise = 1
)
}
\arguments{
\item{id}{a character vector to use as an id for this participant}

\item{intercept}{default value of attitude at t = 0}

\item{slope}{default 1st order regression coefficient}

\item{curvature}{default 2nd order regression coefficient}

\item{noise}{default SD of noise added to attitude at each timepoint}

\item{gender}{a factor string for this participant Boy | Girl}

\item{ses}{a factor string for this participant Low | Medium | High}

\item{from}{first time point}

\item{to}{last time point}

\item{gender_intercept_offset}{Constant that is added to the noise calculations for Girls}

\item{gender_intercept_noise}{Multiplier that is applied to the noise component of the intercept for Girls}

\item{gender_slope_offset}{Constant that is added to the slope calculations for Girls}

\item{gender_slope_noise}{Multiplier that is applied to the noise component of the slope calculations for Girls}

\item{medium_intercept_offset}{Constant that is added to the noise calculations for medium ses participant}

\item{medium_intercept_noise}{Multiplier that is applied to the noise component of the intercept for medium ses participant}

\item{medium_slope_offset}{Constant that is added to the slope calculations for medium ses participant}

\item{medium_slope_noise}{Multiplier that is applied to the noise component of the slope calculations for medium ses participant}

\item{high_intercept_offset}{Constant that is added to the noise calculations for high ses participant}

\item{high_intercept_noise}{Multiplier that is applied to the noise component of the intercept for high ses participant}

\item{high_slope_offset}{Constant that is added to the slope calculations for high ses participant}

\item{high_slope_noise}{Multiplier that is applied to the noise component of the slope calculations for high ses participant}
}
\value{
A tibble of the form |Timepoint|id|gender|ses|attitude|
}
\description{
This function is only called internally
}
\examples{
generate_attitude_sequence('a')

}
