% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_set.R
\name{generate_data_set}
\alias{generate_data_set}
\title{Generate a synthetic dataframe for a 2x2x3 longitudinal design}
\usage{
generate_data_set(
  participants,
  groups = c("control", "intervention"),
  balance = c(1, 1),
  intercept_difference = c(0),
  slope_difference = c(1),
  curvature_difference = c(0),
  noise_difference_multiplier = c(1),
  gender_slope_difference = c(1),
  gender_slope_noise_multiplier = c(1),
  medium_slope_difference = c(0),
  medium_slope_noise_multiplier = c(1),
  high_slope_difference = c(0),
  high_slope_noise_muliplier = c(1),
  ...
)
}
\arguments{
\item{participants}{Number of participants to generate in the data set}

\item{groups}{column vector of names for groups}

\item{balance}{column vector, of same length as groups, with the relative number of participants per group. If the sum of balance is not an integer divisor of participants, then participants will be increased until this condition is met.}

\item{intercept_difference}{how much should be added (in scale points) to the value of initial intercept for group 2 (intervention) etc compared to group 1 (control)? This vector must have length 1 less than groups.}

\item{slope_difference}{How much should be added (in scale points per time period) to the slope of the intervention groups compared to the control group? This vector must have length 1 less than groups.}

\item{curvature_difference}{How much should be added (in scale points per time period squared) to the curvature of the intervention groups compared to the control group? This vector must have length 1 less than groups.}

\item{noise_difference_multiplier}{What multiplier should be attached to the intervention groups compared to the control groups? i.e. how much noisier should the the intervention group be than the control group? This vector must have length 1 less than groups.}

\item{gender_slope_difference}{How much should be added to the Girls group for each intervention compared to the Boys group for the intervention in scale points per time period? i.e. how much more should the intervention affect Girls than Boys? This vector must have length 1 less than groups.}

\item{gender_slope_noise_multiplier}{What multiplier should be used to increase the noise of the girls intervention group compared to the Boys intervention group? i.e. how much noisier should the girls intervention group be than the boys? This vector must have length 1 less than groups.}

\item{medium_slope_difference}{How much should be added to the Medium SES group for each intervention compared to the Low SES group for the intervention in scale points per time period? i.e. how much more should the intervention affect Medium group than the Low group? This vector must have length 1 less than groups.}

\item{medium_slope_noise_multiplier}{What multiplier should be used to increase the noise of the Medium SES intervention group compared to the Low SES intervention group? i.e. how much noisier should the Medium intervention group be than the Low group? This vector must have length 1 less than groups.}

\item{high_slope_difference}{How much should be added to the High SES group for each intervention compared to the Low SES group for the intervention in scale points per time period? i.e. how much more should the intervention affect High group than the Low group? This vector must have length 1 less than groups.}

\item{high_slope_noise_muliplier}{What multiplier should be used to increase the noise of the High SES intervention group compared to the Low SES intervention group? i.e. how much noisier should the High intervention group be than the Low group? This vector must have length 1 less than groups.}

\item{...}{any other parameters to pass to \link{generate_attitude_sequence}}
}
\value{
A long and tidy tibble
}
\description{
This function can be called to generate a semi-deterministic synthetic data set for a 2 Group{Control,Intervention} × 2 Gender{Boy, Girl} × 3 SES{Low,Medium,High} design. The purpose of this is to produce data for testing the other functions in timefitteR.
}
\examples{
x<-generate_data_set()
}
