% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getConsoleInput}
\alias{getConsoleInput}
\title{An Internal helper function to collect and process console input}
\usage{
getConsoleInput(
  prompt,
  check = TRUE,
  valid = NULL,
  invalid = NULL,
  caseSensitive = FALSE,
  msg = NULL,
  type = "any",
  rtn = "string"
)
}
\arguments{
\item{prompt}{A string containing a message to be displayed}

\item{check}{A logical indicating if the input should be checked. If FALSE a character string will be returned irrespective of the value of rtn}

\item{valid}{A list containing valid responses}

\item{invalid}{A list containing invalid responses.}

\item{caseSensitive}{A logical. TRUE means that input will not be passed to toupper() before comparison and return.}

\item{msg}{An optional message to print to the console before the prompt}

\item{type}{one of c("any","character","numeric","param_list") to define acceptable input types}

\item{rtn}{one of c("string","factor","param_list") to define expected return type}
}
\value{
an object defined by type
}
\description{
Displays a console prompt. Checks if input is in valid list (if supplied) else is not in invalid list and returns an object of type rtn.
}
\examples{
x <- getConsoleInput("Do you like ice cream?", valid = c("Yes","No"), type="string")

}
