% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getInput}
\alias{getInput}
\title{An Internal helper function to collect console input}
\usage{
getInput(prompt, check = TRUE, caseSensitive = FALSE, msg = NULL, type = "any")
}
\arguments{
\item{prompt}{A string containing a message to be displayed}

\item{check}{A logical indicating if the input type should be checked. If FALSE a character string will be returned irrespective of the value of rtn}

\item{caseSensitive}{A logical. TRUE means that input will not be passed to toupper() before comparison and return.}

\item{msg}{An optional message to print to the console before the prompt}

\item{type}{one of c("any","character","numeric","list") to define acceptable input types}
}
\value{
an object defined by type
}
\description{
Displays a console prompt. Checks if input is of a valid type and returns an object of type
}
\examples{
x <- getInput("Do you like ice cream?", type="string")

}
