% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timefitteR_Data_Model.R
\name{new_timefitteR_Data_Model}
\alias{new_timefitteR_Data_Model}
\title{Create New TimefitteR Data Model}
\usage{
new_timefitteR_Data_Model(
  time,
  conditions = factor(),
  predictors = list(),
  base_params = timefitteR_Param_List(),
  condition_params = list(),
  predictor_params = list(),
  from = 0L,
  length = 10L,
  step = 1L
)
}
\arguments{
\item{time}{Name of the variable to be used to represent timepoints}

\item{conditions}{A factor variable defining comparison groups e.g., factor(c("Control","Intervention), levels = c("Control","Intervention"))}

\item{predictors}{A list of n factors defining the variables to be used as predictors in exploring the model}

\item{base_params}{A timefitteR Param List that defines the bases values used to generate the data set for the \verb{[1]} levels of each of the n predictor factors}

\item{condition_params}{A list of timefitter_Param_Lists that define the additional parameters for conditions other than the base level}

\item{predictor_params}{A list of lists of timefitter_Param_Lists that define the additional parameters for conditions other than the base level}

\item{from}{The value of the first time point (usually 0)}

\item{length}{The number of time point steps (default 10)}

\item{step}{The step size between time points (usually 1)}
}
\value{
a timefitteR_Data_Model class
}
\description{
The timefitter functions to create synthetic data sets expect to find a certain defined model. This set of functions builds that model. The model general form is:
Model:
-->vars
-->timevariable
-->time_params
-->from
-->length
-->step
-->conditions
-->predictors
-->predictor 1
-->predictor 2
....
-->predictor 1 level 2
-->timefitteR_Param_List
-->predictor 1 level 3 ...
-->predictor 2
--> ...
}
\examples{
y<-new_timefitteR_Data_Model(Timer,conditions = factor(c("Control","Intervention","test")), predictors = list("Gender" = factor(c("Boy","Girl")),"SES" = factor(c("Low","High"),levels = c("Low","High"))), condition_params = list("x"=timefitteR_Param_List(1,2,3,4,5,6),"y"=timefitteR_Param_List(4,5,6,7,8,9)))
y

}
