% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_curves.R
\name{print.compare_curves}
\alias{print.compare_curves}
\title{Print method for compare_curves objects}
\usage{
\method{print}{compare_curves}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"compare_curves"}.}

\item{...}{Unused.}
}
\description{
Print method for compare_curves objects
}
