% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_curve.R
\name{print.fit_curve}
\alias{print.fit_curve}
\title{Print method for fit_curve objects}
\usage{
\method{print}{fit_curve}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"fit_curve"}.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print method for fit_curve objects
}
