% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_curves.R
\name{print.summary_compare_curves}
\alias{print.summary_compare_curves}
\title{Print method for summary_compare_curves objects}
\usage{
\method{print}{summary_compare_curves}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary_compare_curves"}.}

\item{...}{Unused.}
}
\description{
Print method for summary_compare_curves objects
}
