% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_curve.R
\name{print.summary_fit_curve}
\alias{print.summary_fit_curve}
\title{Print method for summary_fit_curve objects}
\usage{
\method{print}{summary_fit_curve}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary_fit_curve"}.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print method for summary_fit_curve objects
}
