% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_curves.R
\name{summary.compare_curves}
\alias{summary.compare_curves}
\title{Summary method for compare_curves objects}
\usage{
\method{summary}{compare_curves}(object, rank_by = c("auto", "aicc", "aic", "bic"), ...)
}
\arguments{
\item{object}{An object of class \code{"compare_curves"}.}

\item{rank_by}{Which criterion to rank by. One of \code{"aicc"}, \code{"aic"}, \code{"bic"}.
Default chooses the first available among these in that order.}

\item{...}{Unused.}
}
\description{
Summary method for compare_curves objects
}
