% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_curve.R
\name{summary.fit_curve}
\alias{summary.fit_curve}
\title{Summary method for fit_curve objects}
\usage{
\method{summary}{fit_curve}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"fit_curve"}.}

\item{...}{Unused.}
}
\value{
An object of class \code{"summary_fit_curve"} containing a richer summary
of the fitted curve, including (when available) full coefficient statistics.
}
\description{
Summary method for fit_curve objects
}
