% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timefitteR_coerce_x_numeric.R
\name{timefitteR_coerce_x_numeric}
\alias{timefitteR_coerce_x_numeric}
\title{coerce_x_numeric}
\usage{
timefitteR_coerce_x_numeric(x, prefix = NULL, suffix = NULL)
}
\arguments{
\item{x}{The input vector usually taken from a dataframe}

\item{prefix}{An optional string that can be removed from the start of the values in x to get to a number}

\item{suffix}{An optional string that can be removed from the end of the values in x to get a number}
}
\value{
A vector of the same length as x coerced to numeric type or an error.
}
\description{
This function is used \strong{internally} to coerce a column of a dataframe into a numeric format suitable for time fitting.
}
\details{
While time data should generally be organised into order by hand as part of the data cleaning and preparation steps, it is sometimes necessary to do this on the fly.This function is able to take a column vector \code{x} usually from a dataframe in a number of formats and manipulate this in a predictable way to return an equivalent numeric vector.
}
\examples{
x <- as.factor(letters[seq_len(10)])
timefitteR_coerce_x_numeric(x)

x <- as.character(paste0("Time",sprintf("\%02d", sample(1:10, 50, replace = TRUE))))
timefitteR_coerce_x_numeric(x, prefix = "Time")

}
