% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimefitteR_Params_List.R
\name{timefitteR_Param_List}
\alias{timefitteR_Param_List}
\title{Create a New timefitteR_Param_List Object}
\usage{
timefitteR_Param_List(
  intercept_offset = 0,
  intercept_noise = 1,
  slope_offset = 0,
  slope_noise = 1,
  curvature_offset = 0,
  curvature_noise = 1
)
}
\arguments{
\item{intercept_offset}{Numeric to control the intercept of the series for this grouping}

\item{intercept_noise}{numeric multiplier to apply to the standard noise level for the intercept for this grouping.}

\item{slope_offset}{numeric to control the slope of the series for this grouping.}

\item{slope_noise}{numeric multiplier to apply to the standard noise level for the slope for this grouping.}

\item{curvature_offset}{numeric to control the curvature of the series for this grouping.}

\item{curvature_noise}{numeric multiplier to apply to the standard noise level for the curvature for this grouping.}
}
\value{
timefitteR_Param_List
}
\description{
This is used to create compliant parameter lists for data models. Default values add no variation.
}
\examples{
x <- timefitteR_Param_List()
}
