% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timefitteR_Params_List.R
\name{validate_timefitteR_Param_List}
\alias{validate_timefitteR_Param_List}
\title{Validate a timefitteR_Param_List object}
\usage{
validate_timefitteR_Param_List(x)
}
\arguments{
\item{x}{a timefitteR_Param_List object}
}
\value{
a timefitteR_Param_List object or error
}
\description{
Validate a timefitteR_Param_List object
}
\examples{
x <- new_timefitteR_Param_list(0, 1, 0, 1, 0, 1)
validate_timefitteR_Param_List(x)
}
