% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timefitteR_Data_Model.R
\name{clean_param_list}
\alias{clean_param_list}
\title{Clean TimefitteR Parameter Lists}
\usage{
clean_param_list(p, l, n)
}
\arguments{
\item{p}{A list of timefitteR_Param_List objects}

\item{l}{A vector containing the levels of the factor}

\item{n}{A  string containing the name of the factor}
}
\value{
A list
}
\description{
A function for internal use only
}
